combining polygonizations in one command
